requirements.dev.txt
-r requirements.txt
mypy==0.470
flake8==3.2.1
pytest-django==3.1.2
pytest==3.0.6
